Flatpicker bug fix, check before delete todo added

+++ src/components/list/controller/controlListElement.js
@@ -8,6 +8,9 @@ export default function controlListElement(
 ) {
 	// DeleteButton
 	deleteButton.addEventListener('click', () => {
+		const listTitle = list.title.toUpperCase();
+		const deleteCheck = masterController.deleteCheck(listTitle);
+		if (!deleteCheck) return;
 		masterController.deleteList(list);
 	});
